+++ R/mst.R
@@ -3,6 +3,17 @@
 #'
 #' @param x numeric vector of x values
 #' @param y numeric vector of y values
+#' @param outlier_rm logical; if TRUE, iteratively trim large MST edges
+#' @param binner an optional function that bins the x and y vectors prior
+#' to triangulation
+#'  Can be:
+#'   \itemize{
+#'     \item `NULL`: no binning (use raw points)
+#'     \item `"hex"` : hexagonal binning following the procedure in the
+#'     graph-theoretic scagnostics paper (start 40x40, halve
+#'              until <= 250 nonempty cells)
+#'     \item a function: user-defined binner
+#'     }
 #' @return A "numeric" object that gives the plot's stringy score.
 #'
 #' @examples
@@ -19,19 +30,26 @@
 #'   sc_stringy(anscombe$x4, anscombe$y4)
 #'
 #' @export
-sc_stringy <- function(x, y) UseMethod("sc_stringy")
+sc_stringy <- function(x, y, outlier_rm = FALSE, binner = NULL) UseMethod("sc_stringy")
 
 #' @rdname sc_stringy
 #' @export
-sc_stringy.default <- function(x, y){
+sc_stringy.default <- function(x, y, outlier_rm = FALSE, binner = NULL){
   #input: x and y are vectors
-  sc <- scree(x, y)
+  sc <- scree(x, y, outlier_rm = outlier_rm, binner = binner)
+  if (is.null(sc$del)) {
+    dc <- sc_dcor(x,y)
+    if (dc >= 1 - 1e-8) {
+      return(1)
+    }
+
+  }
   sc_stringy.scree(sc)
 }
 
 #' @rdname sc_stringy
 #' @export
-sc_stringy.scree <- function(x, y=NULL) {
+sc_stringy.scree <- function(x, y=NULL, outlier_rm = FALSE, binner = NULL) {
   #input: x is a scree, no y
   mst <- gen_mst(x$del, x$weights)
   sc_stringy.igraph(mst)
@@ -39,7 +57,7 @@ sc_stringy.scree <- function(x, y=NULL) {
 
 #' @rdname sc_stringy
 #' @export
-sc_stringy.igraph <- function(x, y=NULL) {
+sc_stringy.igraph <- function(x, y=NULL, outlier_rm = FALSE, binner = NULL) {
   #input: x is the MST igraph object
   vertex_counts <- igraph::degree(x)
   sum(vertex_counts == 2) / (length(vertex_counts) - sum(vertex_counts == 1))
@@ -47,19 +65,26 @@ sc_stringy.igraph <- function(x, y=NULL) {
 
 #' @rdname sc_stringy
 #' @export
-sc_stringy2 <- function(x, y) UseMethod("sc_stringy2")
+sc_stringy2 <- function(x, y, outlier_rm = FALSE, binner = NULL) UseMethod("sc_stringy2")
 
 #' @rdname sc_stringy
 #' @export
-sc_stringy2.default <- function(x, y){
+sc_stringy2.default <- function(x, y, outlier_rm = FALSE, binner = NULL){
   #input: x and y are vectors
-  sc <- scree(x, y)
+  sc <- scree(x, y, outlier_rm = outlier_rm, binner = binner)
+  if (is.null(sc$del)) {
+    dc <- sc_dcor(x,y)
+    if (dc >= 1 - 1e-8) {
+      return(1)
+    }
+
+  }
   sc_stringy2.scree(sc)
 }
 
 #' @rdname sc_stringy
 #' @export
-sc_stringy2.scree <- function(x, y=NULL) {
+sc_stringy2.scree <- function(x, y=NULL, outlier_rm = FALSE, binner = NULL) {
   #input: x is a scree, no y
   mst <- gen_mst(x$del, x$weights)
   sc_stringy2.igraph(mst)
@@ -67,7 +92,7 @@ sc_stringy2.scree <- function(x, y=NULL) {
 
 #' @rdname sc_stringy
 #' @export
-sc_stringy2.igraph <- function(x, y=NULL) {
+sc_stringy2.igraph <- function(x, y=NULL, outlier_rm = FALSE, binner = NULL) {
   #input: x is the MST igraph object
   diameter <- igraph::get_diameter(x)
   length(diameter) / (length(x) - 1)
@@ -77,6 +102,17 @@ sc_stringy2.igraph <- function(x, y=NULL) {
 #'
 #' @param x numeric vector of x values
 #' @param y numeric vector of y values
+#' @param outlier_rm logical; if TRUE, iteratively trim large MST edges
+#' @param binner an optional function that bins the x and y vectors prior
+#' to triangulation
+#'  Can be:
+#'   \itemize{
+#'     \item `NULL`: no binning (use raw points)
+#'     \item `"hex"` : hexagonal binning following the procedure in the
+#'     graph-theoretic scagnostics paper (start 40x40, halve
+#'              until <= 250 nonempty cells)
+#'     \item a function: user-defined binner
+#'     }
 #' @return A "numeric" object that gives the plot's striated score.
 #'
 #' @examples
@@ -90,18 +126,18 @@ sc_stringy2.igraph <- function(x, y=NULL) {
 #'   sc_striated(anscombe$x2, anscombe$y2)
 #'
 #' @export
-sc_striated <- function(x, y) UseMethod("sc_striated")
+sc_striated <- function(x, y, outlier_rm = FALSE, binner = NULL) UseMethod("sc_striated")
 
 #' @rdname sc_striated
 #' @export
-sc_striated.default <- function(x, y){
-  sc <- scree(x, y)
+sc_striated.default <- function(x, y, outlier_rm = FALSE, binner = NULL){
+  sc <- scree(x, y, outlier_rm = outlier_rm, binner = binner)
   sc_striated.scree(sc)
 }
 
 #' @rdname sc_striated
 #' @export
-sc_striated.scree <- function(x, y=NULL) {
+sc_striated.scree <- function(x, y=NULL, outlier_rm = FALSE, binner = NULL) {
   #input: x is scree
   mst <- gen_mst(x$del, x$weights)
   sc_striated.igraph(mst, x)
@@ -110,7 +146,7 @@ sc_striated.scree <- function(x, y=NULL) {
 
 #' @rdname sc_striated
 #' @export
-sc_striated.igraph <- function(x, y){
+sc_striated.igraph <- function(x, y, outlier_rm = FALSE, binner = NULL){
   #input: x is MST, y is scree
   vertex_counts <- igraph::degree(x)
   angs <- which(vertex_counts==2)
@@ -129,6 +165,17 @@ sc_striated.igraph <- function(x, y){
 #'
 #' @param x numeric vector of x values
 #' @param y numeric vector of y values
+#' @param outlier_rm logical; if TRUE, iteratively trim large MST edges
+#' @param binner an optional function that bins the x and y vectors prior
+#' to triangulation
+#'  Can be:
+#'   \itemize{
+#'     \item `NULL`: no binning (use raw points)
+#'     \item `"hex"` : hexagonal binning following the procedure in the
+#'     graph-theoretic scagnostics paper (start 40x40, halve
+#'              until <= 250 nonempty cells)
+#'     \item a function: user-defined binner
+#'     }
 #' @return A "numeric" object that gives the plot's clumpy score.
 #'
 #' @examples
@@ -141,18 +188,18 @@ sc_striated.igraph <- function(x, y){
 #'   sc_clumpy(datasaurus_dozen_wide$away_x, datasaurus_dozen_wide$away_y)
 #'
 #' @export
-sc_clumpy <- function(x, y) UseMethod("sc_clumpy")
+sc_clumpy <- function(x, y, outlier_rm = FALSE, binner = NULL) UseMethod("sc_clumpy")
 
 #' @rdname sc_clumpy
 #' @export
-sc_clumpy.default <- function(x, y){
-  sc <- scree(x, y)
+sc_clumpy.default <- function(x, y, outlier_rm = FALSE, binner = NULL){
+  sc <- scree(x, y, outlier_rm = outlier_rm, binner = binner)
   sc_clumpy.scree(sc)
 }
 
 #' @rdname sc_clumpy
 #' @export
-sc_clumpy.scree <- function(x, y=NULL) {
+sc_clumpy.scree <- function(x, y=NULL, outlier_rm = FALSE, binner = NULL) {
   #input: x is a scree
   mst <- gen_mst(x$del, x$weights)
   sc_clumpy.igraph(mst,x)
@@ -161,7 +208,7 @@ sc_clumpy.scree <- function(x, y=NULL) {
 
 #' @rdname sc_clumpy
 #' @export
-sc_clumpy.igraph <- function(x, y){
+sc_clumpy.igraph <- function(x, y, outlier_rm = FALSE, binner = NULL){
   #input: x is the MST, y is the scree
   #lower triangular matrix
   mstmat <- twomstmat(x,y)$lowertri
@@ -245,6 +292,17 @@ sc_clumpy.igraph <- function(x, y){
 #'
 #' @param x numeric vector of x values
 #' @param y numeric vector of y values
+#' @param outlier_rm logical; if TRUE, iteratively trim large MST edges
+#' @param binner an optional function that bins the x and y vectors prior
+#' to triangulation
+#'  Can be:
+#'   \itemize{
+#'     \item `NULL`: no binning (use raw points)
+#'     \item `"hex"` : hexagonal binning following the procedure in the
+#'     graph-theoretic scagnostics paper (start 40x40, halve
+#'              until <= 250 nonempty cells)
+#'     \item a function: user-defined binner
+#'     }
 #' @return A "numeric" object that gives the plot's sparse score.
 #'
 #' @examples
@@ -259,18 +317,18 @@ sc_clumpy.igraph <- function(x, y){
 #'   sc_sparse(datasaurus_dozen_wide$dino_x, datasaurus_dozen_wide$dino_y)
 #'
 #' @export
-sc_sparse <- function(x, y) UseMethod("sc_sparse")
+sc_sparse <- function(x, y, outlier_rm = FALSE, binner = NULL) UseMethod("sc_sparse")
 
 #' @rdname sc_sparse
 #' @export
-sc_sparse.default <- function(x, y){
-  sc <- scree(x, y)
+sc_sparse.default <- function(x, y, outlier_rm = FALSE, binner = NULL){
+  sc <- scree(x, y, outlier_rm = outlier_rm, binner = binner)
   sc_sparse.scree(sc)
 }
 
 #' @rdname sc_sparse
 #' @export
-sc_sparse.scree <- function(x, y=NULL) {
+sc_sparse.scree <- function(x, y=NULL, outlier_rm = FALSE, binner = NULL) {
   #generate vector of MST edges
   mst <- gen_mst(x$del, x$weights)
   sc_sparse.igraph(mst,x)
@@ -279,7 +337,7 @@ sc_sparse.scree <- function(x, y=NULL) {
 
 #' @rdname sc_sparse
 #' @export
-sc_sparse.igraph <- function(x, y){
+sc_sparse.igraph <- function(x, y, outlier_rm = FALSE, binner = NULL){
   #input: x is MST, y is scree
   mstmat <- twomstmat(x,y)$lowertri
   edges <- mstmat[which(mstmat>0)]
@@ -295,6 +353,17 @@ sc_sparse.igraph <- function(x, y){
 #'
 #' @param x numeric vector of x values
 #' @param y numeric vector of y values
+#' @param outlier_rm logical; if TRUE, iteratively trim large MST edges
+#' @param binner an optional function that bins the x and y vectors prior
+#' to triangulation
+#'  Can be:
+#'   \itemize{
+#'     \item `NULL`: no binning (use raw points)
+#'     \item `"hex"` : hexagonal binning following the procedure in the
+#'     graph-theoretic scagnostics paper (start 40x40, halve
+#'              until <= 250 nonempty cells)
+#'     \item a function: user-defined binner
+#'     }
 #' @return A "numeric" object that gives the plot's skewed score.
 #'
 #' @examples
@@ -310,18 +379,18 @@ sc_sparse.igraph <- function(x, y){
 #'   sc_skewed(datasaurus_dozen_wide$x_shape_x, datasaurus_dozen_wide$x_shape_y)
 #'
 #' @export
-sc_skewed <- function(x, y) UseMethod("sc_skewed")
+sc_skewed <- function(x, y, outlier_rm = FALSE, binner = NULL) UseMethod("sc_skewed")
 
 #' @rdname sc_skewed
 #' @export
-sc_skewed.default <- function(x, y){
-  sc <- scree(x, y)
+sc_skewed.default <- function(x, y, outlier_rm = FALSE, binner = NULL){
+  sc <- scree(x, y, outlier_rm = outlier_rm, binner = binner)
   sc_skewed.scree(sc)
 }
 
 #' @rdname sc_skewed
 #' @export
-sc_skewed.scree <- function(x, y=NULL) {
+sc_skewed.scree <- function(x, y=NULL, outlier_rm = FALSE, binner = NULL) {
   #generate vector of MST edges
   mst <- gen_mst(x$del, x$weights)
   sc_skewed.igraph(mst, x)
@@ -329,7 +398,7 @@ sc_skewed.scree <- function(x, y=NULL) {
 
 #' @rdname sc_skewed
 #' @export
-sc_skewed.igraph <- function(x, y){
+sc_skewed.igraph <- function(x, y, outlier_rm = FALSE, binner = NULL){
   mstmat <- twomstmat(x,y)$lowertri
   edges <- mstmat[which(mstmat>0)]
 
@@ -352,6 +421,17 @@ sc_skewed.igraph <- function(x, y){
 #'
 #' @param x numeric vector of x values
 #' @param y numeric vector of y values
+#' @param outlier_rm logical; if TRUE, iteratively trim large MST edges
+#' @param binner an optional function that bins the x and y vectors prior
+#' to triangulation
+#'  Can be:
+#'   \itemize{
+#'     \item `NULL`: no binning (use raw points)
+#'     \item `"hex"` : hexagonal binning following the procedure in the
+#'     graph-theoretic scagnostics paper (start 40x40, halve
+#'              until <= 250 nonempty cells)
+#'     \item a function: user-defined binner
+#'     }
 #' @return A "numeric" object that gives the plot's outlying score.
 #'
 #' @examples
@@ -366,18 +446,18 @@ sc_skewed.igraph <- function(x, y){
 #'   sc_outlying(datasaurus_dozen_wide$h_lines_x, datasaurus_dozen_wide$h_lines_y)
 #'
 #' @export
-sc_outlying <- function(x, y) UseMethod("sc_outlying")
+sc_outlying <- function(x, y, outlier_rm = FALSE, binner = NULL) UseMethod("sc_outlying")
 
 #' @rdname sc_outlying
 #' @export
-sc_outlying.default <- function(x, y){
-  sc <- scree(x, y)
+sc_outlying.default <- function(x, y, outlier_rm = FALSE, binner = NULL){
+  sc <- scree(x, y, outlier_rm = outlier_rm, binner = binner)
   sc_outlying.scree(sc)
 }
 
 #' @rdname sc_outlying
 #' @export
-sc_outlying.scree <- function(x, y=NULL) {
+sc_outlying.scree <- function(x, y=NULL, outlier_rm = FALSE, binner = NULL) {
   #generate vector of MST edges
   mst <- gen_mst(x$del, x$weights)
   sc_outlying.igraph(mst, x)
@@ -385,7 +465,7 @@ sc_outlying.scree <- function(x, y=NULL) {
 
 #' @rdname sc_outlying
 #' @export
-sc_outlying.igraph <- function(x, y){
+sc_outlying.igraph <- function(x, y, outlier_rm = FALSE, binner = NULL){
   #input: x orig mst (mymst) and y scree object
   #output: outlying mst value